requirements.txt
numpy
scipy
librosa
python-sofa